Create utility class for log related functions
