+++ konduit-serving-core/src/main/java/ai/konduit/serving/configprovider/KonduitServingMain.java
@@ -22,6 +22,7 @@
 
 package ai.konduit.serving.configprovider;
 
+import ai.konduit.serving.util.LogUtils;
 import ai.konduit.serving.verticles.base.BaseRoutableVerticle;
 import ai.konduit.serving.verticles.inference.InferenceVerticle;
 import ch.qos.logback.classic.Level;
@@ -137,7 +138,7 @@ public class KonduitServingMain {
 
     private void deployVerticle(KonduitServingNodeConfigurer konduitServingNodeConfigurer, Vertx vertx) {
         if(konduitServingNodeConfigurer.getInferenceConfiguration().getServingConfig().isCreateLoggingEndpoints()) {
-            setFileAppenderIfNeeded();
+            LogUtils.setFileAppenderIfNeeded();
         }
 
         vertx.deployVerticle(konduitServingNodeConfigurer.getVerticleClassName(), konduitServingNodeConfigurer.getDeploymentOptions(), handler -> {
@@ -164,42 +165,6 @@ public class KonduitServingMain {
                 }
             }
         });
-    }
-
-    public void setFileAppenderIfNeeded() {
-        File previousLogsFile = PipelineRouteDefiner.getLogsFile();
-
-        String konduitServingLogDirFromEnv = System.getenv("KONDUIT_SERVING_LOG_DIR");
-        File newLogsFile = Paths.get(Strings.isNullOrEmpty(konduitServingLogDirFromEnv) ?
-                System.getProperty("user.dir") : konduitServingLogDirFromEnv, "main.log").toFile();
-
-        if(!newLogsFile.equals(previousLogsFile)) {
-            ch.qos.logback.classic.Logger rootLogger = (ch.qos.logback.classic.Logger)
-                    org.slf4j.LoggerFactory.getLogger(ch.qos.logback.classic.Logger.ROOT_LOGGER_NAME);
-
-            LoggerContext context = (LoggerContext) org.slf4j.LoggerFactory.getILoggerFactory();
-
-            rootLogger.setLevel(Level.ALL);
-
-            if(previousLogsFile != null) {
-                rootLogger.detachAppender("FILE");
-            }
-
-            FileAppender<ILoggingEvent> fileAppender = new FileAppender<>();
-            fileAppender.setName("FILE");
-            fileAppender.setFile(newLogsFile.getAbsolutePath());
-            fileAppender.setContext(context);
-
-            PatternLayoutEncoder patternLayoutEncoder = new PatternLayoutEncoder();
-            patternLayoutEncoder.setPattern("%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n");
-            patternLayoutEncoder.setContext(context);
-            patternLayoutEncoder.start();
-
-            fileAppender.setEncoder(patternLayoutEncoder);
-            fileAppender.start();
-
-            rootLogger.addAppender(fileAppender);
-        }
     }
 
     @FunctionalInterface
